requirements.txt
flask
python-dotenv
openai
aiohttp
requests
langchain
chromadb
pypdf
sentence-transformers